++Map Preset
The map spawn is done, it is somewhat premade and somewhat random so....
the best of both worlds I suppose. Anyway I'm going to spruce it up a
lil' bit. Oh btw the buildings are 1x1 sorry if that isn't what everyone
intended its a lot easier to get it to work if it is 1x1 but  we can
edit it later if we need to for now it should work.