+Generate Map Object
Ok so I woke up this morning and checked the code and realized that for
some reason the update didn't commit so I had to redo and recommit it
all so here you go. Also I wasn't sure where you wanted the object to be
generated so I made it in the map class (some real Oedipus time machine
shit there) if that's not where you wanted it go ahead and move it.